+++ app/deserializers/forest_liana/resource_deserializer.rb
@@ -32,7 +32,12 @@ module ForestLiana
     def extract_attributes_serialize
       attributes_serialized.each do |attribute, serializer|
         value = @params[:data][:attributes][attribute]
-        @attributes[attribute] = JSON::parse(value)
+        begin
+          @attributes[attribute] = value.nil? ? nil : JSON::parse(value)
+        rescue
+          message = "Bad format for '#{attribute}' attribute value."
+          raise ForestLiana::Errors::SerializeAttributeBadFormat.new(message)
+        end
       end
     end