+++ app/services/forest_liana/belongs_to_updater.rb
@@ -1,18 +1,27 @@
 module ForestLiana
   class BelongsToUpdater
+    attr_accessor :errors
+
     def initialize(resource, association, params)
       @resource = resource
       @association = association
       @params = params
       @data = params['data']
+      @errors = nil
     end
 
     def perform
-      @record = @resource.find(@params[:id])
-      new_value = @association.klass.find(@data[:id]) if @data && @data[:id]
-      @record.send("#{@association.name}=", new_value)
+      begin
+        @record = @resource.find(@params[:id])
+        new_value = @association.klass.find(@data[:id]) if @data && @data[:id]
+        @record.send("#{@association.name}=", new_value)
 
-      @record.save()
+        @record.save
+      rescue ActiveRecord::SerializationTypeMismatch => exception
+        @errors = [{ detail: exception.message }]
+      rescue => exception
+        @errors = [{ detail: exception.message }]
+      end
     end
   end
 end