+++ test/dummy/db/migrate/20170614141921_create_serialize_field.rb
@@ -1,0 +1,7 @@
+class CreateSerializeField < ActiveRecord::Migration
+  def change
+    create_table :serialize_fields do |t|
+      t.string :field
+    end
+  end
+end